Refactor: Split broker profile scraping logic
Refactor the broker profile scraping function into smaller, more maintainable files. This includes splitting the extraction logic into separate utility files to improve code organization and readability.
