+++ supabase/config.toml
@@ -12,4 +12,3 @@ verify_jwt = false
 
 [functions.scrape-broker-profile]
 verify_jwt = false
-
